refactor: KOB-116 — split orchestrator core modules
Extract engine routing, chat-tab lifecycle, task running, and worktree coordination out of the Orchestrator aggregate. Keep the public Orchestrator API as a thin facade while reducing core.ts below 1000 lines.
